lib: sms: Add new SMS module
The SMS module allows to have multiple SMS clients in the system. It
acts as a central SMS client and dispatch SMS data to multiple
listeners. This module acknowledge SMS on the behalf on each listeners.

Multiple listener can register and get notified when a SMS is received.
SMS are read using AT commands, using the shared AT socket.

Signed-off-by: Christopher Métrailler <christopher.metrailler@nordicsemi.no>